Move the All variable away from gadget/main.c and into libgadget/init.c
This is an effort to stop link-time optimization breaking every time
sizeof(All) changes
