+++ gadget/main.c
@@ -7,8 +7,8 @@
 #include <math.h>
 #include <gsl/gsl_math.h>
 #include <gsl/gsl_errno.h>
+#include <omp.h>
 
-#include <libgadget/allvars.h>
 #include <libgadget/slotsmanager.h>
 #include <libgadget/partmanager.h>
 
@@ -74,7 +74,9 @@ int main(int argc, char **argv)
 
     tamalloc_init();
 
-    read_parameter_file(argv[1]);	/* ... read in parameters for this run */
+    int ShowBacktrace;
+    double MaxMemSizePerNode;
+    read_parameter_file(argv[1], &ShowBacktrace, &MaxMemSizePerNode);	/* ... read in parameters for this run */
 
     int RestartFlag, RestartSnapNum;
 
@@ -98,23 +100,23 @@ int main(int argc, char **argv)
     }
 
     if(RestartFlag == 1) {
-        RestartSnapNum = find_last_snapnum(All.OutputDir);
-        message(0, "Last Snapshot number is %d.\n", RestartSnapNum);
+        /* Last snapshot will be detected in begrun*/
+        RestartSnapNum = -2;
     }
 
     /*Set up GSL so it gives a proper MPI termination*/
     gsl_set_error_handler(gsl_handler);
 
     /*Initialize the memory manager*/
-    mymalloc_init(All.MaxMemSizePerNode);
+    mymalloc_init(MaxMemSizePerNode);
 
     /* Make sure memory has finished initialising on all ranks before doing more.
      * This may improve stability */
     MPI_Barrier(MPI_COMM_WORLD);
 
-    init_endrun(All.ShowBacktrace);
+    init_endrun(ShowBacktrace);
 
-    begrun(RestartSnapNum);
+    RestartSnapNum = begrun(RestartSnapNum);
 
     switch(RestartFlag) {
         case 3:
