+++ gadget/params.h
@@ -1,2 +1,4 @@
-
-void read_parameter_file(char *fname);
+#ifndef __GADGET_PARAMS_H
+#define __GADGET_PARAMS_H
+void read_parameter_file(char *fname, int * ShowBacktrace, double * MaxMemSizePerNode);
+#endif
