+++ libgadget/run.h
@@ -1,7 +1,7 @@
 #ifndef RUN_H
 #define RUN_H
 
-void begrun(int RestartSnapNum);
+int begrun(int RestartSnapNum);
 
 void run(int RestartSnapNum);
 void runtests(int RestartSnapNum);
